refactor profile conformance 'checker' to 'helper'
